src/requirement.txt
exchangelib==5.1.0
ibm-db==3.1.4
oauthlib==3.2.2
python-decouple==3.8
requests==2.31.0
requests-oauthlib==1.3.1
